BoundaryValueDiffEqAscher

Gauss Legendre collocation methods with Ascher's error control adaptivity and mesh refinement routines. To be able to access the solvers in BoundaryValueDiffEqFIRK, you must first install them use the Julia package manager:

using Pkg
Pkg.add("BoundaryValueDiffEqAscher")

solve(prob::BVProblem, alg, dt; kwargs...)
solve(prob::TwoPointBVProblem, alg, dt; kwargs...)

Full List of Methods

Ascher1: 1 stage Gauss Legendre collocation method with Ascher's error control adaptivity and mesh refinement.

Ascher2: 2 stage Gauss Legendre collocation method with Ascher's error control adaptivity and mesh refinement.

Ascher3: 3 stage Gauss Legendre collocation method with Ascher's error control adaptivity and mesh refinement.

Ascher4: 4 stage Gauss Legendre collocation method with Ascher's error control adaptivity and mesh refinement.

Ascher5: 5 stage Gauss Legendre collocation method with Ascher's error control adaptivity and mesh refinement.

Ascher6: 6 stage Gauss Legendre collocation method with Ascher's error control adaptivity and mesh refinement.

Ascher7: 7 stage Gauss Legendre collocation method with Ascher's error control adaptivity and mesh refinement.

Detailed Solvers Explanation

BoundaryValueDiffEqAscher.Ascher1 — Type

Ascher1(; nlsolve = NewtonRaphson(), max_num_subintervals = 3000)

1th stage Gauss Legendre collocation methods with adaptivity adapted from Ascher's implementation.

Keyword Arguments

nlsolve: Internal Nonlinear solver. Any solver which conforms to the SciML NonlinearProblem interface can be used. Note that any autodiff argument for the solver will be ignored and a custom jacobian algorithm will be used.

max_num_subintervals: Number of maximal subintervals, default as 3000.

zeta: side condition points, should always be provided.

Note

For type-stability, the chunksizes for ForwardDiff ADTypes in BVPJacobianAlgorithm must be provided.

References

@article{Ascher1994CollocationSF,
    title={Collocation Software for Boundary Value Differential-Algebraic Equations},
    author={Uri M. Ascher and Raymond J. Spiteri},
    journal={SIAM J. Sci. Comput.},
    year={1994},
    volume={15},
    pages={938-952},
    url={https://api.semanticscholar.org/CorpusID:10597070}
}

@article{Ascher1979ACS,
    title={A collocation solver for mixed order systems of boundary value problems},
    author={Uri M. Ascher and J. Christiansen and Robert D. Russell},
    journal={Mathematics of Computation},
    year={1979},
    volume={33},
    pages={659-679},
    url={https://api.semanticscholar.org/CorpusID:121729124}
}

BoundaryValueDiffEqAscher.Ascher2 — Type

Ascher2(; nlsolve = NewtonRaphson(), max_num_subintervals = 3000)

2th stage Gauss Legendre collocation methods with adaptivity adapted from Ascher's implementation.

Keyword Arguments

nlsolve: Internal Nonlinear solver. Any solver which conforms to the SciML NonlinearProblem interface can be used. Note that any autodiff argument for the solver will be ignored and a custom jacobian algorithm will be used.

max_num_subintervals: Number of maximal subintervals, default as 3000.

zeta: side condition points, should always be provided.

Note

For type-stability, the chunksizes for ForwardDiff ADTypes in BVPJacobianAlgorithm must be provided.

References

@article{Ascher1994CollocationSF,
    title={Collocation Software for Boundary Value Differential-Algebraic Equations},
    author={Uri M. Ascher and Raymond J. Spiteri},
    journal={SIAM J. Sci. Comput.},
    year={1994},
    volume={15},
    pages={938-952},
    url={https://api.semanticscholar.org/CorpusID:10597070}
}

@article{Ascher1979ACS,
    title={A collocation solver for mixed order systems of boundary value problems},
    author={Uri M. Ascher and J. Christiansen and Robert D. Russell},
    journal={Mathematics of Computation},
    year={1979},
    volume={33},
    pages={659-679},
    url={https://api.semanticscholar.org/CorpusID:121729124}
}

BoundaryValueDiffEqAscher.Ascher3 — Type

Ascher3(; nlsolve = NewtonRaphson(), max_num_subintervals = 3000)

3th stage Gauss Legendre collocation methods with adaptivity adapted from Ascher's implementation.

Keyword Arguments

nlsolve: Internal Nonlinear solver. Any solver which conforms to the SciML NonlinearProblem interface can be used. Note that any autodiff argument for the solver will be ignored and a custom jacobian algorithm will be used.

max_num_subintervals: Number of maximal subintervals, default as 3000.

zeta: side condition points, should always be provided.

Note

For type-stability, the chunksizes for ForwardDiff ADTypes in BVPJacobianAlgorithm must be provided.

References

@article{Ascher1994CollocationSF,
    title={Collocation Software for Boundary Value Differential-Algebraic Equations},
    author={Uri M. Ascher and Raymond J. Spiteri},
    journal={SIAM J. Sci. Comput.},
    year={1994},
    volume={15},
    pages={938-952},
    url={https://api.semanticscholar.org/CorpusID:10597070}
}

@article{Ascher1979ACS,
    title={A collocation solver for mixed order systems of boundary value problems},
    author={Uri M. Ascher and J. Christiansen and Robert D. Russell},
    journal={Mathematics of Computation},
    year={1979},
    volume={33},
    pages={659-679},
    url={https://api.semanticscholar.org/CorpusID:121729124}
}

BoundaryValueDiffEqAscher.Ascher4 — Type

Ascher4(; nlsolve = NewtonRaphson(), max_num_subintervals = 3000)

4th stage Gauss Legendre collocation methods with adaptivity adapted from Ascher's implementation.

Keyword Arguments

nlsolve: Internal Nonlinear solver. Any solver which conforms to the SciML NonlinearProblem interface can be used. Note that any autodiff argument for the solver will be ignored and a custom jacobian algorithm will be used.

max_num_subintervals: Number of maximal subintervals, default as 3000.

zeta: side condition points, should always be provided.

Note

For type-stability, the chunksizes for ForwardDiff ADTypes in BVPJacobianAlgorithm must be provided.

References

@article{Ascher1994CollocationSF,
    title={Collocation Software for Boundary Value Differential-Algebraic Equations},
    author={Uri M. Ascher and Raymond J. Spiteri},
    journal={SIAM J. Sci. Comput.},
    year={1994},
    volume={15},
    pages={938-952},
    url={https://api.semanticscholar.org/CorpusID:10597070}
}

@article{Ascher1979ACS,
    title={A collocation solver for mixed order systems of boundary value problems},
    author={Uri M. Ascher and J. Christiansen and Robert D. Russell},
    journal={Mathematics of Computation},
    year={1979},
    volume={33},
    pages={659-679},
    url={https://api.semanticscholar.org/CorpusID:121729124}
}

BoundaryValueDiffEqAscher.Ascher5 — Type

Ascher5(; nlsolve = NewtonRaphson(), max_num_subintervals = 3000)

5th stage Gauss Legendre collocation methods with adaptivity adapted from Ascher's implementation.

Keyword Arguments

nlsolve: Internal Nonlinear solver. Any solver which conforms to the SciML NonlinearProblem interface can be used. Note that any autodiff argument for the solver will be ignored and a custom jacobian algorithm will be used.

max_num_subintervals: Number of maximal subintervals, default as 3000.

zeta: side condition points, should always be provided.

Note

For type-stability, the chunksizes for ForwardDiff ADTypes in BVPJacobianAlgorithm must be provided.

References

@article{Ascher1994CollocationSF,
    title={Collocation Software for Boundary Value Differential-Algebraic Equations},
    author={Uri M. Ascher and Raymond J. Spiteri},
    journal={SIAM J. Sci. Comput.},
    year={1994},
    volume={15},
    pages={938-952},
    url={https://api.semanticscholar.org/CorpusID:10597070}
}

@article{Ascher1979ACS,
    title={A collocation solver for mixed order systems of boundary value problems},
    author={Uri M. Ascher and J. Christiansen and Robert D. Russell},
    journal={Mathematics of Computation},
    year={1979},
    volume={33},
    pages={659-679},
    url={https://api.semanticscholar.org/CorpusID:121729124}
}

BoundaryValueDiffEqAscher.Ascher6 — Type

Ascher6(; nlsolve = NewtonRaphson(), max_num_subintervals = 3000)

6th stage Gauss Legendre collocation methods with adaptivity adapted from Ascher's implementation.

Keyword Arguments

nlsolve: Internal Nonlinear solver. Any solver which conforms to the SciML NonlinearProblem interface can be used. Note that any autodiff argument for the solver will be ignored and a custom jacobian algorithm will be used.

max_num_subintervals: Number of maximal subintervals, default as 3000.

zeta: side condition points, should always be provided.

Note

For type-stability, the chunksizes for ForwardDiff ADTypes in BVPJacobianAlgorithm must be provided.

References

@article{Ascher1994CollocationSF,
    title={Collocation Software for Boundary Value Differential-Algebraic Equations},
    author={Uri M. Ascher and Raymond J. Spiteri},
    journal={SIAM J. Sci. Comput.},
    year={1994},
    volume={15},
    pages={938-952},
    url={https://api.semanticscholar.org/CorpusID:10597070}
}

@article{Ascher1979ACS,
    title={A collocation solver for mixed order systems of boundary value problems},
    author={Uri M. Ascher and J. Christiansen and Robert D. Russell},
    journal={Mathematics of Computation},
    year={1979},
    volume={33},
    pages={659-679},
    url={https://api.semanticscholar.org/CorpusID:121729124}
}

BoundaryValueDiffEqAscher.Ascher7 — Type

Ascher7(; nlsolve = NewtonRaphson(), max_num_subintervals = 3000)

7th stage Gauss Legendre collocation methods with adaptivity adapted from Ascher's implementation.

Keyword Arguments

nlsolve: Internal Nonlinear solver. Any solver which conforms to the SciML NonlinearProblem interface can be used. Note that any autodiff argument for the solver will be ignored and a custom jacobian algorithm will be used.

max_num_subintervals: Number of maximal subintervals, default as 3000.

zeta: side condition points, should always be provided.

Note

For type-stability, the chunksizes for ForwardDiff ADTypes in BVPJacobianAlgorithm must be provided.

References

@article{Ascher1994CollocationSF,
    title={Collocation Software for Boundary Value Differential-Algebraic Equations},
    author={Uri M. Ascher and Raymond J. Spiteri},
    journal={SIAM J. Sci. Comput.},
    year={1994},
    volume={15},
    pages={938-952},
    url={https://api.semanticscholar.org/CorpusID:10597070}
}

@article{Ascher1979ACS,
    title={A collocation solver for mixed order systems of boundary value problems},
    author={Uri M. Ascher and J. Christiansen and Robert D. Russell},
    journal={Mathematics of Computation},
    year={1979},
    volume={33},
    pages={659-679},
    url={https://api.semanticscholar.org/CorpusID:121729124}
}